+++ tergite_autocalibration/lib/base/node.py
@@ -14,16 +14,13 @@
 # that they have been altered from the originals.
 
 import abc
-import json
-import threading
-import time
 from collections.abc import Iterable
 from pathlib import Path
-from typing import List, Optional, Tuple
+from typing import List, Literal, Optional, Tuple
 
 import matplotlib
 import numpy as np
-import tqdm
+import quantify_scheduler.backends.qblox.constants as constants
 import xarray
 from colorama import Fore, Style
 from colorama import init as colorama_init
@@ -31,29 +28,20 @@ from quantify_scheduler.backends import SerialCompiler
 from quantify_scheduler.device_under_test.quantum_device import QuantumDevice
 from quantify_scheduler.instrument_coordinator.instrument_coordinator import (
     CompiledSchedule,
+    InstrumentCoordinator,
 )
-from quantify_scheduler.json_utils import SchedulerJSONEncoder, pathlib
 
 from tergite_autocalibration.config import settings
-from tergite_autocalibration.config.settings import HARDWARE_CONFIG, REDIS_CONNECTION
 from tergite_autocalibration.lib.base.analysis import BaseNodeAnalysis
 from tergite_autocalibration.lib.base.measurement import BaseMeasurement
-from tergite_autocalibration.lib.utils.redis import (
-    load_redis_config,
-    load_redis_config_coupler,
-)
-from tergite_autocalibration.tools.mss.convert import structured_redis_storage
+from tergite_autocalibration.lib.utils.device import DeviceConfiguration
+from tergite_autocalibration.lib.utils.schedule_execution import execute_schedule
 from tergite_autocalibration.utils.dataset_utils import configure_dataset, save_dataset
 from tergite_autocalibration.utils.dto.enums import MeasurementMode
-from tergite_autocalibration.utils.extended_coupler_edge import CompositeSquareEdge
-from tergite_autocalibration.utils.extended_transmon_element import ExtendedTransmon
 from tergite_autocalibration.utils.logger.tac_logger import logger
 
 colorama_init()
 
-
-with open(HARDWARE_CONFIG) as hw:
-    hw_config = json.load(hw)
 
 matplotlib.use(settings.PLOTTING_BACKEND)
 
@@ -69,19 +57,21 @@ class BaseNode(abc.ABC):
         self.all_qubits = all_qubits
         self.node_dictionary = node_dictionary
         self.backup = False
-        self.type = "simple_sweep"  # TODO better as Enum type
         self.qubit_state = 0  # can be 0 or 1 or 2
         self.plots_per_qubit = 1  # can be 0 or 1 or 2
-
-        self.coupler: Optional[List[str]]
-
-        self.lab_instr_coordinator = None
+        self.couplers: Optional[List[str] | None] = None
+        self.lab_instr_coordinator: InstrumentCoordinator
+        self.measured_elements: Literal["Single_Qubits", "Couplers"] = "Single_Qubits"
 
         self.schedule_samplespace = {}
         self.external_samplespace = {}
+
+        # These may be modified while the node runs
+        self.outer_schedule_samplespace = {}
         self.initial_schedule_samplespace = {}
-        self.schedule_keywords = {}
         self.reduced_external_samplespace = {}
+        self.loops = None
+        self.schedule_keywords = {}
 
         self.samplespace = self.schedule_samplespace | self.external_samplespace
 
@@ -96,9 +86,42 @@ class BaseNode(abc.ABC):
                 "Quantities of Interest are missing from the node implementation"
             )
 
+        # NOTE: In the future this will be problematic.
+        # Having the device creation in the init will prohibit concurrent
+        # initialization of two different nodes
+        self.device_manager = DeviceConfiguration(self.all_qubits, self.couplers)
+        self.device = self.device_manager.configure_device(self.name)
+
+    def measure_node(self, cluster_status) -> xarray.Dataset:
+        result_dataset = xarray.Dataset()
+        """
+        To be implemented by the Classes that define the Node Type:
+        ScheduleNode or ExternalParameterNode
+        """
+        return result_dataset
+
     def pre_measurement_operation(self):
         """
         To be implemented by the child measurement nodes
+        """
+        pass
+
+    def initial_operation(self):
+        """
+        To be implemented by the child measurement nodes.
+        This is called before the execution of each and every iteration
+        of the samples of the external samplespace.
+        See coupler_spectroscopy for examples.
+        """
+        pass
+
+    def final_operation(self):
+        """
+        To be implemented by the child measurement nodes.
+        This is called after ALL the iteration samples of the
+        external samplespace have been executed.
+        e.g. set back the dc_current to 0 in coupler_spectroscopy.
+        See coupler_spectroscopy for examples.
         """
         pass
 
@@ -106,14 +129,8 @@ class BaseNode(abc.ABC):
     def dimensions(self) -> list:
         """
         array of dimensions used for raw dataset reshaping
-        in utills/dataset_utils.py. some nodes have peculiar dimensions
-        e.g. randomized benchmarking and need dimension definition in their class
         """
         schedule_settable_quantities = self.schedule_samplespace.keys()
-
-        # no schedule_samplespace applies on to sc_qubit_spectroscopy
-        if len(list(schedule_settable_quantities)) == 0:
-            return [1]
 
         # keeping the first element, ASSUMING that all settable elements
         # have the same dimensions on their samplespace
@@ -130,245 +147,60 @@ class BaseNode(abc.ABC):
                 settable_values = np.array([settable_values])
             dimensions.append(len(settable_values))
 
-        if self.external_samplespace != {} and self.initial_schedule_samplespace == {}:
-            dimensions = dimensions + [1]
+        if self.loops is not None:
+            dimensions.append(self.loops)
         return dimensions
 
-    @property
-    def external_dimensions(self) -> list:
-        """
-        array of dimensions used for raw dataset reshaping
-        in utills/dataset_utils.py. some nodes have peculiar dimensions
-        e.g. randomized benchmarking and need dimension definition in their class
-        """
-        external_settable_quantities = self.external_samplespace.keys()
-
-        # keeping the first element, ASSUMING that all settable elements
-        # have the same dimensions on their samplespace
-        # i.e. all qubits have the same number of ro frequency samples in readout spectroscopy
-        first_settable = list(external_settable_quantities)[0]
-        measured_elements = self.external_samplespace[first_settable].keys()
-        first_element = list(measured_elements)[0]
-
-        dimensions = []
-        if len(dimensions) > 1:
-            raise NotImplementedError("Multidimensional External Samplespace")
-        for quantity in external_settable_quantities:
-            dimensions.append(len(self.external_samplespace[quantity][first_element]))
-        return dimensions
-
-    def calibrate(self, data_path: Path, lab_ic, cluster_status):
+    def calibrate(self, data_path: Path, cluster_status):
         if cluster_status != MeasurementMode.re_analyse:
-            self.run_measurement(data_path, lab_ic, cluster_status)
+            result_dataset = self.measure_node(cluster_status)
+            self.device_manager.save_serial_device(self.name, self.device, data_path)
+            # After the measurement free the device resources
+            save_dataset(result_dataset, self.name, data_path)
+        self.device_manager.close_device()
         self.post_process(data_path)
         logger.info("analysis completed")
 
-    def run_measurement(self, data_path: Path, lab_ic, cluster_status):
-        compiled_schedule = self.precompile(data_path)
+    def precompile(self, schedule_samplespace: dict) -> CompiledSchedule:
+        constants.GRID_TIME_TOLERANCE_TIME = 5e-2
 
-        if self.external_samplespace == {}:
-            """
-            This correspond to simple cluster schedules
-            """
-            result_dataset = self.measure_node(
-                compiled_schedule,
-                lab_ic,
-                data_path,
-                cluster_status=cluster_status,
-            )
-        else:
-            pre_measurement_operation = self.pre_measurement_operation
-
-            # node.external_dimensions is defined in the node_base
-            iterations = self.external_dimensions[0]
-
-            result_dataset = xarray.Dataset()
-
-            # example of external_samplespace:
-            # external_samplespace = {
-            #       'cw_frequencies': {
-            #          'q1': np.array(4.0e9, 4.1e9, 4.2e9),
-            #          'q2': np.array(4.5e9, 4.6e9, 4.7e9),
-            #        }
-            # }
-
-            # e.g. 'cw_frequencies':
-            external_settable = list(self.external_samplespace.keys())[0]
-
-            for current_iteration in range(iterations):
-                reduced_external_samplespace = {}
-                qubit_values = {}
-                # elements may refer to qubits or couplers
-                elements = self.external_samplespace[external_settable].keys()
-                for element in elements:
-                    qubit_specific_values = self.external_samplespace[
-                        external_settable
-                    ][element]
-                    external_value = qubit_specific_values[current_iteration]
-                    qubit_values[element] = external_value
-
-                # example of reduced_external_samplespace:
-                # reduced_external_samplespace = {
-                #     'cw_frequencies': {
-                #          'q1': np.array(4.2e9),
-                #          'q2': np.array(4.7e9),
-                #     }
-                # }
-                reduced_external_samplespace[external_settable] = qubit_values
-                self.reduced_external_samplespace = reduced_external_samplespace
-                pre_measurement_operation(
-                    reduced_ext_space=reduced_external_samplespace
-                )
-
-                ds = self.measure_node(
-                    compiled_schedule,
-                    lab_ic,
-                    data_path,
-                    cluster_status,
-                    measurement=(current_iteration, iterations),
-                )
-                result_dataset = xarray.merge([result_dataset, ds])
-        logger.info("measurement completed")
-
-    def precompile(
-        self, data_path: Path, bin_mode: str = None, repetitions: int = None
-    ):
+        # TODO: put 'tof' out of its misery
         if self.name == "tof":
             return None, 1
-        qubits = self.all_qubits
 
-        # backup old parameter values
-        # TODO:
-        if self.backup:
-            fields = self.redis_field
-            for field in fields:
-                field_backup = field + "_backup"
-                for qubit in qubits:
-                    key = f"transmons:{qubit}"
-                    if field in REDIS_CONNECTION.hgetall(key).keys():
-                        value = REDIS_CONNECTION.hget(key, field)
-                        REDIS_CONNECTION.hset(key, field_backup, value)
-                        REDIS_CONNECTION.hset(key, field, "nan")
-                        structured_redis_storage(field_backup, qubit.strip("q"), value)
-                        REDIS_CONNECTION.hset(key, field, "nan")
-                        structured_redis_storage(field, qubit.strip("q"), None)
-                if getattr(self, "coupler", None) is not None:
-                    couplers = self.coupler
-                    for coupler in couplers:
-                        key = f"couplers:{coupler}"
-                        if field in REDIS_CONNECTION.hgetall(key).keys():
-                            value = REDIS_CONNECTION.hget(key, field)
-                            REDIS_CONNECTION.hset(key, field_backup, value)
-                            structured_redis_storage(field_backup, coupler, value)
-                            REDIS_CONNECTION.hset(key, field, "nan")
-                            structured_redis_storage(key, coupler, value)
+        transmons = self.device_manager.transmons
 
-        device = QuantumDevice(f"Loki_{self.name}")
-        device.hardware_config(hw_config)
-
-        transmons = {}
-        for channel, qubit in enumerate(qubits):
-            transmon = ExtendedTransmon(qubit)
-            transmon = load_redis_config(transmon, channel)
-            device.add_element(transmon)
-            transmons[qubit] = transmon
-
-        # Creating coupler edge
-        # bus_list = [ [qubits[i],qubits[i+1]] for i in range(len(qubits)-1) ]
-        if hasattr(self, "edges"):
-            couplers = self.edges
-            edges = {}
-            for bus in couplers:
-                control, target = bus.split(sep="_")
-                coupler = CompositeSquareEdge(control, target)
-                load_redis_config_coupler(coupler)
-                device.add_edge(coupler)
-                edges[bus] = coupler
-
-        if hasattr(self, "edges") or self.name in [
-            "cz_chevron",
-            "cz_calibration",
-            "cz_calibration_ssro",
-            "cz_calibration_swap_ssro",
-            "cz_dynamic_phase",
-            "cz_dynamic_phase_swap",
-            "cz_parametrisation_fix_duration",
-            "reset_chevron",
-            "reset_calibration_ssro",
-            "tqg_randomized_benchmarking",
-            "tqg_randomized_benchmarking_interleaved",
-        ]:
-            coupler = self.coupler
-            node_class = self.measurement_obj(transmons, edges, self.qubit_state)
+        if self.measured_elements == "Couplers":
+            edges = self.device_manager.edges
+            node_class = self.measurement_obj(transmons, edges)
         else:
-            node_class = self.measurement_obj(transmons, self.qubit_state)
-        if self.name in [
-            "ro_amplitude_three_state_optimization",
-            "cz_calibration_ssro",
-            "cz_calibration_swap_ssro",
-            "reset_calibration_ssro",
-        ]:
-            device.cfg_sched_repetitions(1)  # for single-shot readout
-        if bin_mode is not None:
-            node_class.set_bin_mode(bin_mode)
+            node_class = self.measurement_obj(transmons)
+        schedule = node_class.schedule_function(
+            **schedule_samplespace, **self.schedule_keywords
+        )
 
-        schedule_function = node_class.schedule_function
-
+        # TODO: Probably the compiler desn't need to be created every time self.precompile() is called.
         compiler = SerialCompiler(name=f"{self.name}_compiler")
 
-        schedule_samplespace = self.schedule_samplespace
-        external_samplespace = self.external_samplespace
-        schedule_keywords = self.schedule_keywords
-
-        schedule = schedule_function(**schedule_samplespace, **schedule_keywords)
-        compilation_config = device.generate_compilation_config()
-
-        # save_serial_device(device, data_path)
-
-        # create a transmon with the same name but with updated config
-        # get the transmon template in dictionary form
-        serialized_device = json.dumps(device, cls=SchedulerJSONEncoder)
-        decoded_device = json.loads(serialized_device)
-        serial_device = {}
-        for element, element_config in decoded_device["data"]["elements"].items():
-            serial_config = json.loads(element_config)
-            serial_device[element] = serial_config
-
-        data_path.mkdir(parents=True, exist_ok=True)
-        with open(f"{data_path}/{self.name}.json", "w") as f:
-            json.dump(serial_device, f, indent=4)
-
-        device.close()
-
-        # after the compilation_config is acquired, free the transmon resources
-        for extended_transmon in transmons.values():
-            extended_transmon.close()
-        if hasattr(self, "edges"):
-            for extended_edge in edges.values():
-                extended_edge.close()
-
+        compilation_config = self.device.generate_compilation_config()
         logger.info("Starting Compiling")
-
         compiled_schedule = compiler.compile(
             schedule=schedule, config=compilation_config
         )
 
         return compiled_schedule
 
-    def measure_node(
+    def measure_compiled_schedule(
         self,
         compiled_schedule: CompiledSchedule,
-        lab_ic,
-        data_path: pathlib.Path,
         cluster_status=MeasurementMode.real,
         measurement: Tuple[int, int] = (1, 1),
-    ) -> None:
+    ) -> xarray.Dataset:
         """
         Execute a measurement for a node and save the resulting dataset.
 
         Args:
             compiled_schedule (CompiledSchedule): The compiled schedule to execute.
-            lab_ic: The lab instrument controller.
             data_path (pathlib.Path): Path where the dataset will be saved.
             measurement (tuple): Tuple of (current_measurement, total_measurements).
         """
@@ -376,13 +208,16 @@ class BaseNode(abc.ABC):
         schedule_duration = self._calculate_schedule_duration(compiled_schedule)
         self._print_measurement_info(schedule_duration, measurement)
 
-        raw_dataset = self.execute_schedule(
-            compiled_schedule, lab_ic, schedule_duration, cluster_status
+        raw_dataset = execute_schedule(
+            compiled_schedule,
+            schedule_duration,
+            self.lab_instr_coordinator,
+            cluster_status,
         )
         result_dataset = configure_dataset(raw_dataset, self)
-        save_dataset(result_dataset, self.name, data_path)
 
         logger.info("Finished measurement")
+        return result_dataset
 
     def _calculate_schedule_duration(
         self, compiled_schedule: CompiledSchedule
@@ -406,44 +241,6 @@ class BaseNode(abc.ABC):
         print(
             f"schedule_duration = {Fore.CYAN}{Style.BRIGHT}{message}{Style.RESET_ALL}"
         )
-
-    def execute_schedule(
-        self,
-        compiled_schedule: CompiledSchedule,
-        lab_ic,
-        schedule_duration: float,
-        cluster_status=None,
-    ) -> xarray.Dataset:
-        # TODO: Could move to helper function, because is static
-
-        logger.info("Starting measurement")
-
-        def run_measurement() -> None:
-            lab_ic.prepare(compiled_schedule)
-            lab_ic.start()
-            lab_ic.wait_done(timeout_sec=3600)
-
-        def display_progress():
-            steps = int(schedule_duration * 5)
-            if cluster_status == MeasurementMode.real:
-                progress_sleep = 0.2
-            for _ in tqdm.tqdm(
-                range(steps), desc=compiled_schedule.name, colour="blue"
-            ):
-                time.sleep(progress_sleep)
-
-        thread_tqdm = threading.Thread(target=display_progress)
-        thread_tqdm.start()
-        thread_lab = threading.Thread(target=run_measurement)
-        thread_lab.start()
-        thread_lab.join()
-        thread_tqdm.join()
-
-        raw_dataset: xarray.Dataset = lab_ic.retrieve_acquisition()
-        lab_ic.stop()
-        logger.info("Raw dataset acquired")
-
-        return raw_dataset
 
     def post_process(self, data_path: Path):
         analysis_kwargs = getattr(self, "analysis_kwargs", dict())