+++ tergite_autocalibration/lib/nodes/coupler/reset_chevron/measurement.py
@@ -11,41 +11,25 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-"""
-Module containing a schedule class for Ramsey calibration. (1D parameter sweep, for 2D see ramsey_detunings.py)
-"""
 import numpy as np
 from quantify_scheduler import Schedule
 from quantify_scheduler.enums import BinMode
-from quantify_scheduler import Schedule
-from quantify_scheduler.operations.gate_library import Measure, Reset, X90, Rxy, X, CZ
+from quantify_scheduler.operations.gate_library import Measure, Reset, Rxy, X
 from quantify_scheduler.operations.pulse_library import (
-    GaussPulse,
-    SuddenNetZeroPulse,
-    ResetClockPhase,
     IdlePulse,
-    DRAGPulse,
-    SetClockFrequency,
-    NumericalPulse,
-    SoftSquarePulse,
-    SquarePulse,
 )
 from quantify_scheduler.operations.pulse_library import (
     RampPulse,
-    DRAGPulse,
-    SetClockFrequency,
-    NumericalPulse,
-    SoftSquarePulse,
-    SquarePulse,
-    ResetClockPhase,
 )
 from quantify_scheduler.resources import ClockResource
 
-from tergite_autocalibration.utils.extended_coupler_edge import CompositeSquareEdge
+from tergite_autocalibration.config.coupler_config import qubit_types
+from tergite_autocalibration.lib.base.measurement import BaseMeasurement
+from tergite_autocalibration.utils.extended_coupler_edge import (
+    ExtendedCompositeSquareEdge,
+)
+from tergite_autocalibration.utils.extended_gates import Rxy_12
 from tergite_autocalibration.utils.extended_transmon_element import ExtendedTransmon
-from tergite_autocalibration.utils.extended_gates import Measure_RO1, Rxy_12
-from ....base.measurement import BaseMeasurement
-from tergite_autocalibration.config.coupler_config import edge_group, qubit_types
 
 
 class Reset_Chevron_DC(BaseMeasurement):
@@ -54,7 +38,7 @@ class Reset_Chevron_DC(BaseMeasurement):
     def __init__(
         self,
         transmons: dict[str, ExtendedTransmon],
-        couplers: dict[str, CompositeSquareEdge],
+        couplers: dict[str, ExtendedCompositeSquareEdge],
         qubit_state: int = 0,
     ):
         super().__init__(transmons)