+++ tergite_autocalibration/lib/nodes/coupler/spectroscopy/node.py
@@ -11,22 +11,29 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-from pathlib import Path
-
 import numpy as np
-from scipy import optimize as optimize
 
+from tergite_autocalibration.lib.nodes.coupler.spectroscopy.analysis import (
+    CouplerSpectroscopyNodeAnalysis,
+)
+from tergite_autocalibration.lib.nodes.qubit_control.spectroscopy.measurement import (
+    Two_Tones_Multidim,
+)
+from tergite_autocalibration.lib.nodes.readout.resonator_spectroscopy.measurement import (
+    Resonator_Spectroscopy,
+)
 from tergite_autocalibration.utils.dto.enums import MeasurementMode
-from tergite_autocalibration.utils.hardware_utils import SpiDAC
-from tergite_autocalibration.utils.user_input import qubit_samples, resonator_samples
+from tergite_autocalibration.utils.hardware.spi import SpiDAC
 
-from ....base.node import BaseNode
-from ...qubit_control.spectroscopy.measurement import Two_Tones_Multidim
-from ...readout.resonator_spectroscopy.measurement import Resonator_Spectroscopy
-from .analysis import CouplerSpectroscopyNodeAnalysis
+# TODO: check location
+from tergite_autocalibration.utils.user_input import qubit_samples
+
+from tergite_autocalibration.lib.base.external_parameter_node import (
+    ExternalParameterNode,
+)
 
 
-class Coupler_Spectroscopy_Node(BaseNode):
+class Coupler_Spectroscopy_Node(ExternalParameterNode):
     measurement_obj = Two_Tones_Multidim
     analysis_obj = CouplerSpectroscopyNodeAnalysis
     coupler_qois = ["parking_current", "current_range"]
@@ -79,7 +86,7 @@ class Coupler_Spectroscopy_Node(BaseNode):
         self.spi_dac.set_dac_current(self.dac, this_iteration_value)
 
 
-class Coupler_Resonator_Spectroscopy_Node(BaseNode):
+class Coupler_Resonator_Spectroscopy_Node(ExternalParameterNode):
     measurement_obj = Resonator_Spectroscopy
     analysis_obj = CouplerSpectroscopyNodeAnalysis
     coupler_qois = ["resonator_flux_quantum"]