+++ tergite_autocalibration/lib/nodes/readout/resonator_spectroscopy/analysis.py
@@ -90,8 +90,6 @@ class ResonatorSpectroscopyQubitAnalysis(BaseQubitAnalysis):
             )
             # using the min value driectly
             self.min_freq_data = self.frequencies[np.argmin(np.abs(self.s21_values))]
-            # print(fit_Ql)
-            # print(self.min_freq )
             return [self.minimum_freq, fit_Ql, self.min_freq_data]
 
     def plotter(self, ax):