+++ tergite_autocalibration/scripts/calibration_supervisor.py
@@ -18,6 +18,7 @@
 # - Martin Ahindura, 2023
 
 from ipaddress import IPv4Address
+from pathlib import Path
 from typing import List, Union
 
 import toml
@@ -39,7 +40,6 @@ from tergite_autocalibration.lib.utils.graph import filtered_topological_order
 from tergite_autocalibration.lib.utils.node_factory import NodeFactory
 from tergite_autocalibration.utils.dataset_utils import create_node_data_path
 from tergite_autocalibration.utils.dto.enums import DataStatus, MeasurementMode
-from tergite_autocalibration.utils.logger.errors import ClusterNotFoundError
 from tergite_autocalibration.utils.logger.tac_logger import logger
 from tergite_autocalibration.utils.logger.visuals import draw_arrow_chart
 from tergite_autocalibration.utils.redis_utils import (
@@ -82,6 +82,8 @@ class CalibrationSupervisor:
         self.measurement_mode: "MeasurementMode" = measurement_mode
         self.cluster_ip: Union[str, "IPv4Address"] = cluster_ip
         self.cluster_timeout: int = cluster_timeout
+        self.node_name_to_re_analyse = node_name
+        self.data_path = Path(data_path)
 
         # Create objects to communicate with the hardware
         self.cluster: "Cluster" = self._create_cluster()
@@ -92,6 +94,8 @@ class CalibrationSupervisor:
         self.qubits = user_requested_calibration["all_qubits"]
         self.couplers = user_requested_calibration["couplers"]
         self.target_node = user_requested_calibration["target_node"]
+        if self.measurement_mode == MeasurementMode.re_analyse:
+            self.target_node = self.node_name_to_re_analyse
         self.user_samplespace = user_requested_calibration["user_samplespace"]
 
         # Read the device configuration
@@ -156,6 +160,7 @@ class CalibrationSupervisor:
         # TODO: everything which is not in the inspect or calibrate function should go here
         logger.info("Starting System Calibration")
         number_of_qubits = len(self.qubits)
+        print(self.topo_order)
         draw_arrow_chart(f"Qubits: {number_of_qubits}", self.topo_order)
 
         # TODO: check if coupler node status throws error after REDISFLUSHALL
@@ -175,26 +180,11 @@ class CalibrationSupervisor:
             logger.info(f"{calibration_node} node is completed")
 
     def inspect_node(self, node_name: str):
+        # TODO: this function must be split
         logger.info(f"Inspecting node {node_name}")
-
-        node: BaseNode = self.calibration_node_factory.create_node(
-            node_name,
-            self.qubits,
-            couplers=self.couplers,
-            measurement_mode=self.measurement_mode,
-        )
-
-        if node.name in self.user_samplespace:
-            update_to_user_samplespace(node, self.user_samplespace)
-
-        # it's maybe useful to give access to the ic
-        node.lab_instr_coordinator = self.lab_ic
-
         populate_initial_parameters(
             self.transmon_configuration, self.qubits, self.couplers, REDIS_CONNECTION
         )
-        # print(f'{node_name = }')
-        # print(f'{self.couplers = }')
         if node_name in [
             "coupler_spectroscopy",
             "cz_chevron",
@@ -204,6 +194,8 @@ class CalibrationSupervisor:
             "cz_calibration_swap_ssro",
             "cz_dynamic_phase",
             "cz_dynamic_phase_swap",
+            "reset_chevron",
+            "process_tomography_ssro",
             "tqg_randomized_benchmarking",
             "tqg_randomized_benchmarking_interleaved",
         ]:
@@ -211,7 +203,7 @@ class CalibrationSupervisor:
                 REDIS_CONNECTION.hget(f"cs:{coupler}", node_name) == "calibrated"
                 for coupler in self.couplers
             ]
-            # print(f'{coupler_statuses=}')
+
             # node is calibrated only when all couplers have the node calibrated:
             is_node_calibrated = all(coupler_statuses)
         else:
@@ -233,7 +225,6 @@ class CalibrationSupervisor:
 
         # Check Redis if node is calibrated
         status = DataStatus.undefined
-
         if node_name in [
             "coupler_spectroscopy",
             "cz_chevron",
@@ -242,13 +233,14 @@ class CalibrationSupervisor:
             "cz_chevron_experimental",
             "cz_optimize_chevron",
             "cz_chevron_amplitude",
-            "cz_calibration",
             "cz_calibration_ssro",
             "cz_calibration_swap_ssro",
-            "cz_dynamic_phase",
-            "cz_dynamic_phase_swap",
-            "tqg_randomized_benchmarking",
-            "tqg_randomized_benchmarking_interleaved",
+            "cz_dynamic_phase_ssro",
+            "cz_dynamic_phase_swap_ssro",
+            "reset_chevron",
+            "process_tomography_ssro",
+            "tqg_randomized_benchmarking_ssro",
+            "tqg_randomized_benchmarking_interleaved_ssro",
         ]:
             for coupler in self.couplers:
                 # the calibrated, not_calibrated flags may be not necessary,
@@ -274,23 +266,44 @@ class CalibrationSupervisor:
                 else:
                     raise ValueError(f"status: {status}")
 
-        if status == DataStatus.in_spec:
+        if (
+            status == DataStatus.in_spec
+            and self.measurement_mode == MeasurementMode.re_analyse
+            and self.node_name_to_re_analyse != node_name
+        ) or (
+            status == DataStatus.in_spec
+            and self.measurement_mode != MeasurementMode.re_analyse
+        ):
             print(
                 f" \u2714  {Fore.GREEN}{Style.BRIGHT}Node {node_name} in spec{Style.RESET_ALL}"
             )
             return
 
-        elif status == DataStatus.out_of_spec:
+        else:
             print(
                 "\u2691\u2691\u2691 "
                 + f"{Fore.RED}{Style.BRIGHT}Calibration required for Node {node_name}{Style.RESET_ALL}"
             )
+
+            node: BaseNode = self.calibration_node_factory.create_node(
+                node_name,
+                self.qubits,
+                couplers=self.couplers,
+                measurement_mode=self.measurement_mode,
+            )
+            if node.name in self.user_samplespace:
+                update_to_user_samplespace(node, self.user_samplespace)
+            # it's maybe useful to give access to the ic
+            node.lab_instr_coordinator = self.lab_ic
+
             logger.info(f"Calibrating node {node.name}")
             # TODO: This could be in the node initializer
-            data_path = create_node_data_path(node)
-            measurement_result = node.calibrate(
-                data_path, self.lab_ic, self.measurement_mode
-            )
+            if self.measurement_mode == MeasurementMode.re_analyse:
+                data_path = self.data_path
+            else:
+                data_path = create_node_data_path(node)
+
+            measurement_result = node.calibrate(data_path, self.measurement_mode)
 
             # TODO:  develop failure strategies ->
             # if node_calibration_status == DataStatus.out_of_spec: